+++ qtEEGsignals/EEGsignals/thetaSignalDecoder.h
@@ -1,12 +1,12 @@
 #ifndef THETASIGNALDECODER_H
 #define THETASIGNALDECODER_H
-#include "datagramDecoder.h"
+#include "DatagramDecoder.h"
 class ThetaSignalDecoder :	public DatagramDecoder
 {
 public:
 	ThetaSignalDecoder();
 	~ThetaSignalDecoder();
-	long long decodeSignal(const QByteArray &datagram);
+	double decodeSignal(const QByteArray& datagram);
 };
 
 #endif
